requirements.txt
torch
transformers
sentence-transformers
faiss-cpu
streamlit
pandas
numpy
scikit-learn
accelerate